pyproject.toml
[tool.poetry]
name = "clan-sports"
version = "0.2.2"
description = ""
authors = ["codigouranio"]
license = "GPL-3.0"
readme = "README.md"
packages = []


[tool.poetry.dependencies]
python = "^3.11"
black = "^23.12.1"
blinker = "^1.7.0"
click = "^8.1.7"
cssmin = "^0.2.0"
faker = "^21.0.0"
flake8 = "^6.1.0"
flask-assets = "^2.1.0"
flask-sessions = "^0.1.5"
flask = "^3.0.0"
gunicorn = "^21.2.0"
isort = "^5.13.2"
itsdangerous = "^2.1.2"
jinja2 = "^3.1.2"
jsmin = "^3.0.1"
lesscpy = "^0.15.1"
markupsafe = "^2.1.3"
mccabe = "^0.7.0"
mypy-extensions = "^1.0.0"
mypy = "^1.8.0"
packaging = "^23.2"
pathspec = "^0.12.1"
platformdirs = "^4.1.0"
ply = "^3.11"
pycodestyle = "^2.11.1"
pyflakes = "^3.1.0"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
six = "^1.16.0"
sqlalchemy = "^2.0.27"
typing-extensions = "^4.9.0"
webassets = "^2.0"
werkzeug = "^3.0.1"
flask-login = "^0.6.3"
flask-cors = "^4.0.0"
flask-session = "^0.6.0"
redis = "^5.0.3"
dependency-injector = "^4.41.0"
uwsgi = { version = "^2.0.24", markers = "python_version >= '3.11' and sys_platform == 'linux'" }
requests = "^2.31.0"
colorama = "0.2.5"
boto3 = "^1.34.65"
flask-sqlalchemy = "^3.1.1"
alembic = "^1.13.1"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.0.0"
qrcode = "^7.4.2"
pip = "^24.0"
pillow = "^10.3.0"
flask-caching = "^2.3.0"
mnemonic = "^0.21"
bip32utils = "^0.3.post4"
ecdsa = "^0.19.0"
pycryptodome = "^3.20.0"
base58 = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
